# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Question Bank API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions with optional filters
      parameters:
        - name: chapterId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [EASY, MEDIUM, HARD]
        - name: type
          in: query
          schema:
            type: string
            enum: [MCQ, FILLUP]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionListResponse"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question (with optional image upload)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID (with choices)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question with choices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /questions/{id}/image:
    get:
      operationId: getQuestionImage
      tags: [questions]
      summary: Serve question image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Image binary
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: No image
  /questions/{id}/preview:
    get:
      operationId: previewQuestion
      tags: [questions]
      summary: Get full question preview data with base64 images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question preview with base64 images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionPreview"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # CHOICES
  /choices:
    post:
      operationId: createChoice
      tags: [choices]
      summary: Create a choice for a question
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateChoiceBody"
      responses:
        "201":
          description: Created choice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Choice"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Question:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        chapterName:
          type: ["string", "null"]
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        text:
          type: string
        type:
          type: string
          enum: [MCQ, FILLUP]
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        imageUrl:
          type: ["string", "null"]
        imageName:
          type: ["string", "null"]
        imageType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        choiceCount:
          type: ["integer", "null"]
      required:
        - id
        - chapterId
        - text
        - type
        - difficulty
        - createdAt
    QuestionListResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - questions
        - total
        - page
        - limit
        - totalPages
    CreateQuestionBody:
      type: object
      properties:
        chapterId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [MCQ, FILLUP]
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        image:
          type: string
          format: binary
      required:
        - chapterId
        - text
        - type
        - difficulty
    UpdateQuestionBody:
      type: object
      properties:
        chapterId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [MCQ, FILLUP]
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        image:
          type: string
          format: binary
        removeImage:
          type: string
          enum: ["true", "false"]
    Choice:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        text:
          type: string
        imageUrl:
          type: ["string", "null"]
        imageName:
          type: ["string", "null"]
        isCorrect:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - questionId
        - text
        - isCorrect
        - createdAt
    CreateChoiceBody:
      type: object
      properties:
        questionId:
          type: integer
        text:
          type: string
        isCorrect:
          type: boolean
        image:
          type: string
          format: binary
      required:
        - questionId
        - text
        - isCorrect
...
    QuestionDetail:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        chapterName:
          type: ["string", "null"]
        subjectId:
          type: ["integer", "null"]
        subjectName:
          type: ["string", "null"]
        text:
          type: string
        type:
          type: string
          enum: [MCQ, FILLUP]
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        imageUrl:
          type: ["string", "null"]
        imageName:
          type: ["string", "null"]
        imageType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        choices:
          type: array
          items:
            $ref: "#/components/schemas/Choice"
      required:
        - id
        - chapterId
        - text
        - type
        - difficulty
        - createdAt
        - choices
    ChoicePreview:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        isCorrect:
          type: boolean
        imageData:
          type: ["string", "null"]
        imageType:
          type: ["string", "null"]
      required:
        - id
        - text
        - isCorrect
    QuestionPreview:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [MCQ, FILLUP]
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        chapterName:
          type: ["string", "null"]
        subjectName:
          type: ["string", "null"]
        imageData:
          type: ["string", "null"]
        imageType:
          type: ["string", "null"]
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChoicePreview"
      required:
        - id
        - text
        - type
        - difficulty
        - choices
...