# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Question Bank API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBodyDifficulty } from "./createQuestionBodyDifficulty";
import type { CreateQuestionBodyType } from "./createQuestionBodyType";

export interface CreateQuestionBody {
  chapterId: number;
  text: string;
  type: CreateQuestionBodyType;
  difficulty: CreateQuestionBodyDifficulty;
  image?: Blob;
}