# lib/api-zod/src/generated/types/questionListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Question Bank API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Question } from "./question";

export interface QuestionListResponse {
  questions: Question[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}